update schema, api

+++ internal/transport/http/handler/schema_handler.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"be/internal/service"
+	"be/internal/shared/constant"
 	response "be/internal/shared/helper"
 	"be/internal/transport/http/dto"
 
@@ -12,14 +13,38 @@ type SchemaHandler struct {
 	schemaService service.ISchemaService
 }
 
-func NewSchemaService(schemaService service.ISchemaService) *SchemaHandler {
+func NewSchemaHandler(schemaService service.ISchemaService) *SchemaHandler {
 	return &SchemaHandler{schemaService: schemaService}
 }
 
+func (h *SchemaHandler) GetSchemas(c *gin.Context) {
+	schemas, err := h.schemaService.GetSchemas(c.Request.Context())
+	if err != nil {
+		response.RespondError(c, err)
+		return
+	}
+	response.RespondSuccess(c, schemas)
+}
+
+func (h *SchemaHandler) GetSchemaByPublicId(c *gin.Context) {
+	id := c.Param("id")
+	if id == "" {
+		response.RespondError(c, &constant.BadRequest)
+		return
+	}
+	schemas, err := h.schemaService.GetSchemaByPublicId(c.Request.Context(), id)
+	if err != nil {
+		response.RespondError(c, err)
+		return
+	}
+	response.RespondSuccess(c, schemas)
+}
+
 func (h *SchemaHandler) CreateSchema(c *gin.Context) {
-	var request *dto.SchemaCreatedRequestDto
+	var request *dto.SchemaBuilderDto
 	if err := c.ShouldBindJSON(request); err != nil {
 		response.RespondError(c, err)
+		return
 	}
 	schema, err := h.schemaService.CreateSchema(c.Request.Context(), request)
 	if err != nil {
@@ -30,9 +55,14 @@ func (h *SchemaHandler) CreateSchema(c *gin.Context) {
 
 func (h *SchemaHandler) RemoveSchema(c *gin.Context) {
 	id := c.Param("id")
+	if id == "" {
+		response.RespondError(c, &constant.BadRequest)
+		return
+	}
 	err := h.schemaService.RemoveSchema(c, id)
 	if err != nil {
 		response.RespondError(c, err)
+		return
 	}
 	response.RespondSuccess(c, nil)
 }
